AudioEngine:refine thread pool
