+++ cocos/audio/win32/AudioEngine-win32.h
@@ -38,8 +38,6 @@ NS_CC_BEGIN
     namespace experimental{
 #define MAX_AUDIOINSTANCES 32
 
-class AudioEngineThreadPool;
-
 class CC_DLL AudioEngineImpl : public cocos2d::Ref
 {
 public:
@@ -67,8 +65,6 @@ public:
     
 private:
     void _play2d(AudioCache *cache, int audioID);
-    
-    AudioEngineThreadPool* _threadPool;
     
     ALuint _alSources[MAX_AUDIOINSTANCES];
