+++ cocos/audio/winrt/AudioEngine-winrt.cpp
@@ -21,101 +21,13 @@
 #if CC_TARGET_PLATFORM == CC_PLATFORM_WINRT
 
 #include "AudioEngine-winrt.h"
-#include <condition_variable>
-
 
 using namespace cocos2d;
 using namespace cocos2d::experimental;
 
-
-namespace cocos2d {
-    namespace experimental {
-        class AudioEngineThreadPool
-        {
-        public:
-            AudioEngineThreadPool()
-                : _running(true)
-                , _numThread(6)
-            {
-                _threads.reserve(_numThread);
-                _tasks.reserve(_numThread);
-
-                for (int index = 0; index < _numThread; ++index) {
-                    _tasks.push_back(nullptr);
-                    _threads.push_back(std::thread(std::bind(&AudioEngineThreadPool::threadFunc, this, index)));
-                }
-            }
-
-            void addTask(const std::function<void()> &task){
-                _taskMutex.lock();
-                int targetIndex = -1;
-                for (int index = 0; index < _numThread; ++index) {
-                    if (_tasks[index] == nullptr) {
-                        targetIndex = index;
-                        _tasks[index] = task;
-                        break;
-                    }
-                }
-                if (targetIndex == -1) {
-                    _tasks.push_back(task);
-                    _threads.push_back(std::thread(std::bind(&AudioEngineThreadPool::threadFunc, this, _numThread)));
-
-                    _numThread++;
-                }
-                _taskMutex.unlock();
-                _sleepCondition.notify_all();
-            }
-
-            void destroy()
-            {
-                _running = false;
-                _sleepCondition.notify_all();
-
-                for (int index = 0; index < _numThread; ++index) {
-                    _threads[index].join();
-                }
-            }
-        private:
-            bool _running;
-            std::vector<std::thread>  _threads;
-            std::vector< std::function<void()> > _tasks;
-
-            void threadFunc(int index)
-            {
-                while (_running) {
-                    std::function<void()> task = nullptr;
-                    _taskMutex.lock();
-                    task = _tasks[index];
-                    _taskMutex.unlock();
-
-                    if (nullptr == task)
-                    {
-                        std::unique_lock<std::mutex> lk(_sleepMutex);
-                        _sleepCondition.wait(lk);
-                        continue;
-                    }
-
-                    task();
-
-                    _taskMutex.lock();
-                    _tasks[index] = nullptr;
-                    _taskMutex.unlock();
-                }
-            }
-
-            int _numThread;
-
-            std::mutex _taskMutex;
-            std::mutex _sleepMutex;
-            std::condition_variable _sleepCondition;
-        };
-    }
-}
-
 AudioEngineImpl::AudioEngineImpl()
     : _lazyInitLoop(true)
     , _currentAudioID(0)
-    , _threadPool(nullptr)
 {
 
 }
@@ -123,20 +35,11 @@ AudioEngineImpl::AudioEngineImpl()
 AudioEngineImpl::~AudioEngineImpl()
 {
     _audioCaches.clear();
-
-    if (_threadPool) {
-        _threadPool->destroy();
-        delete _threadPool;
-    }
 }
 
 bool AudioEngineImpl::init()
 {
     bool ret = false;
-
-    if (nullptr == _threadPool) {
-        _threadPool = new (std::nothrow) AudioEngineThreadPool();
-    }
 
     ret = true;
     return ret;
@@ -173,7 +76,7 @@ AudioCache* AudioEngineImpl::preload(const std::string& filePath)
 
             std::string fullPath = FileUtils::getInstance()->fullPathForFilename(filePath);
             audioCache->_fileFullPath = fullPath;
-            _threadPool->addTask(std::bind(&AudioCache::readDataTask, audioCache));
+            AudioEngine::addTask(std::bind(&AudioCache::readDataTask, audioCache));
         }
         else {
             audioCache = &it->second;
